requirements.txt
crewai>=0.108.0
crewai-tools>=0.33.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pillow-heif>=0.18.0
sentence-transformers>=3.0.0
